pyproject.toml
[project]
name = "grovescan"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "coverage>=7.10.3",
    "pillow>=11.3.0",
    "psutil>=7.0.0",
    "pyinsane2>=2.0.13",
    "pyinstaller>=6.15.0",
    "pyside6>=6.9.1",
    "pywin32>=311",
    "reportlab>=4.4.3",
    "twain-wia-sane-scanner>=2.0.2",
]
